pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dusa-backend"
version = "1.0.0"
description = "Down Under Stats Backend"
authors = [
    "Alex Ford <alextford1@gmail.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
fastapi = "^0.104.1"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.24.0.post1"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
sqlalchemy-utils = "^0.41.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pre-commit = "^3.5.0"
ruff = "^0.1.6"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"
pytest-factoryboy = "^2.6.0"

[tool.ruff]
line-length = 120

[tool.ruff.lint.isort]
combine-as-imports = true
